switched to shiro for encryption. provided a strtucture around ciphers
