+++ cas-server-core/src/main/java/org/jasig/cas/util/NoOpCipherExecutor.java
@@ -27,15 +27,16 @@ import org.slf4j.LoggerFactory;
  * @author Misagh Moayyed
  * @since 4.1
  */
-public final class NoOpCipherExecutor implements CipherExecutor<String, String> {
+public final class NoOpCipherExecutor extends AbstractCipherExecutor<String, String> {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(NoOpCipherExecutor.class);
 
     /**
-     * Instantiates a new No op cipher executor.
+     * Instantiates a new No-Op cipher executor.
      * Issues a warning on safety.
      */
     public NoOpCipherExecutor() {
+        super(NoOpCipherExecutor.class.getName());
         LOGGER.warn("[{}] does no encryption and may NOT be safe in a production environment. "
                 + "Consider using other choices, such as [{}] that handle encryption, signing and verification of "
                 + "all appropriate values.", this.getClass().getName(), DefaultCipherExecutor.class.getName());
