+++ cas-server-core/src/test/java/org/jasig/cas/util/ShiroCipherExecutorTests.java
@@ -24,17 +24,16 @@ import org.junit.Test;
 import static org.junit.Assert.*;
 
 /**
- * Test cases for {@link JCEBasedCipherExecutor}.
+ * Test cases for {@link ShiroCipherExecutor}.
  * @author Misagh Moayyed
  * @since 4.2
  */
-public class JCEBasedCipherExecutorTests {
+public class ShiroCipherExecutorTests {
 
     @Test
     public void testEncodingDecoding() {
         final String value = "ThisIsATestValueThatIsGoingToBeEncodedAndDecodedAgainAndAgain";
-        final CipherExecutor<byte[], byte[]> cc = new JCEBasedCipherExecutor("1234567890123456",
-                "1234567890123456",
+        final CipherExecutor<byte[], byte[]> cc = new ShiroCipherExecutor("1234567890123456",
                 "szxK-5_eJjs-aUj-64MpUZ-GPPzGLhYPLGl0wrYjYNVAGva2P0lLe6UGKGM7k8dWxsOVGutZWgvmY3l5oVPO3w");
         final byte[] bytes = cc.encode(value.getBytes());
         final byte[] decoded = cc.decode(bytes);
@@ -44,7 +43,7 @@ public class JCEBasedCipherExecutorTests {
     @Test(expected=RuntimeException.class)
     public void testEncodingDecodingBadKeys() {
         final String value = "ThisIsATestValueThatIsGoingToBeEncodedAndDecodedAgainAndAgain";
-        final CipherExecutor<byte[], byte[]> cc = new JCEBasedCipherExecutor("0000", "9999", "1234");
+        final CipherExecutor<byte[], byte[]> cc = new ShiroCipherExecutor("0000", "1234");
         cc.encode(value.getBytes());
     }
 }
